Reorganize tests. Add mocha-snapshot
