+++ packages/lib-react-components/src/components/zooniverse-logo.test.js
@@ -3,16 +3,16 @@
 /* eslint import/no-extraneous-dependencies: ["error", { "devDependencies": true  }] */
 import { mount } from 'enzyme';
 import React from 'react';
-import ZooniverseLogotype from '../src/components/zooniverse-logotype';
+import ZooniverseLogo from './zooniverse-logo';
 
-describe('ZooniverseLogotype', function() {
-  it('mounts ZooniverseLogotype', function() {
-    const wrapper = mount(<ZooniverseLogotype />);
-    expect(wrapper.find(ZooniverseLogotype)).to.have.length(1);
+describe('ZooniverseLogo', function() {
+  it('mounts ZooniverseLogo', function() {
+    const wrapper = mount(<ZooniverseLogo />);
+    expect(wrapper.find('ZooniverseLogo')).to.have.length(1);
   });
 
   it('has a unique aria-labelledby for each instance mounted', function() {
-    const wrapper = mount(<div><ZooniverseLogotype /><ZooniverseLogotype /></div>);
+    const wrapper = mount(<div><ZooniverseLogo /><ZooniverseLogo /></div>);
     const firstSVGlabel = wrapper.find('svg').first().prop('aria-labelledby');
     const secondSVGlabel = wrapper.find('svg').last().prop('aria-labelledby');
     expect(firstSVGlabel).to.not.equal(secondSVGlabel);
